requirements.txt
pbr==3.1.1 # Apache 2.0
python-dotenv==0.7.1 # BSD 3-clause
pyyaml==3.12 # MIT
ingredients.http # MIT